refactor(-): :hammer: move serializers logs dietas
