+++ sme_sigpae_api/dieta_especial/__tests__/test_serializers.py
@@ -1,8 +1,10 @@
 import pytest
 
 from sme_sigpae_api.dieta_especial.api.serializers import (
-    LogQuantidadeDietasAutorizadasRecreioNasFeriasSerializer,
     SolicitacaoDietaEspecialRecreioNasFeriasSerializer,
+)
+from sme_sigpae_api.dieta_especial.logs_models.api.serializers import (
+    LogQuantidadeDietasAutorizadasRecreioNasFeriasSerializer,
 )
 from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadasRecreioNasFerias,
